@@ -1,34 +1,39 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-   <PropertyGroup>
-      <TargetFramework>netcoreapp2.1</TargetFramework>
-   </PropertyGroup>
-   <ItemGroup>
-      <PackageReference Include="Humanizer.Core" Version="2.4.2"/>
-      <PackageReference Include="Microsoft.AspNetCore.App"/>
-      <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.1.3"/>
-      <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="2.1.3"/>
-      <PackageReference Include="OdeToCode.UseNodeModules" Version="1.0.6"/>
-      <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="2.1.1.1"/>
-      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.1.3"/>
-   </ItemGroup>
-   <ItemGroup>
-      <ProjectReference Include="..\NoteWorx.FX\NoteWorx.FX.csproj"/>
-      <ProjectReference Include="..\NoteWorx.Infrastructure.Data\NoteWorx.Infrastructure.Data.csproj"/>
-   </ItemGroup>
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
-   <Target Name="CopyCustomContent" AfterTargets="AfterBuild">
-      <Copy SourceFiles="./Infrastructure/SeedFiles/notes.json" DestinationFolder="$(OutDir)" />
-      <Copy SourceFiles="./Infrastructure/SeedFiles/tags.json" DestinationFolder="$(OutDir)" />
-      <Copy SourceFiles="./Infrastructure/SeedFiles/users.json" DestinationFolder="$(OutDir)" />
-   </Target>
-   <Target Name="CopyCustomContentOnPublish" AfterTargets="Publish">
-      <Copy SourceFiles="./Infrastructure/SeedFiles/notes.json" DestinationFolder="$(PublishDir)" />
-      <Copy SourceFiles="./Infrastructure/SeedFiles/tags.json" DestinationFolder="$(PublishDir)" />
-      <Copy SourceFiles="./Infrastructure/SeedFiles/users.json" DestinationFolder="$(PublishDir)" />
-   </Target>
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <LangVersion>latest</LangVersion>
+    <Nullable>enable</Nullable>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <WarningsAsErrors />
+    <AnalysisMode>Default</AnalysisMode>
+  </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Humanizer.Core" Version="2.9.9" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="5.0.5" />
+    <PackageReference Include="OdeToCode.UseNodeModules" Version="3.0.0" />
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="5.0.5.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.5" />
+  </ItemGroup>
 
-   <Target Name="MyPreCompileTarget" BeforeTargets="Build">
-      <Exec Command="gulp"/>
-   </Target>
+  <ItemGroup>
+    <ProjectReference Include="..\NoteWorx.FX\NoteWorx.FX.csproj" />
+    <ProjectReference Include="..\NoteWorx\NoteWorx.csproj" />
+  </ItemGroup>
+
+  <Target Name="CopyCustomContent" AfterTargets="AfterBuild">
+    <Copy SourceFiles="./Infrastructure/SeedFiles/notes.json" DestinationFolder="$(OutDir)" />
+    <Copy SourceFiles="./Infrastructure/SeedFiles/users.json" DestinationFolder="$(OutDir)" />
+  </Target>
+
+  <Target Name="CopyCustomContentOnPublish" AfterTargets="Publish">
+    <Copy SourceFiles="./Infrastructure/SeedFiles/notes.json" DestinationFolder="$(PublishDir)" />
+    <Copy SourceFiles="./Infrastructure/SeedFiles/users.json" DestinationFolder="$(PublishDir)" />
+  </Target>
+
+  <Target Name="MyPreCompileTarget" BeforeTargets="Build">
+    <Exec Command="gulp" />
+  </Target>
 </Project>
