@@ -1,21 +1,21 @@
-using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.DependencyInjection;
-using NoteWorx.Infrastructure.Data;
-using NoteWorx.Infrastructure.Data.Models;
+using NoteWorx.Identity.Data.Models;
+using NoteWorx.Notes.Data;
 
 namespace NoteWorx.Web.Infrastructure.DependencyInjection
 {
-   internal static partial class AppServiceCollectionExtensions
-   {
-      private static IServiceCollection ConfigureIdentityServices(
-         this IServiceCollection services)
-      {
-         services
-            .AddIdentity<AppUser, AppRole>()
-            .AddEntityFrameworkStores<NoteWorxDbContext>()
-            .AddDefaultTokenProviders();
+    internal static partial class AppServiceCollectionExtensions
+    {
+        private static IServiceCollection ConfigureIdentityServices(
+           this IServiceCollection services)
+        {
+            services
+               .AddIdentity<AppUser, AppRole>()
+               .AddEntityFrameworkStores<NoteWorxDbContext>()
+               .AddDefaultTokenProviders();
 
-         return services;
-      }
-   }
+            return services;
+        }
+    }
 }
