@@ -11,7 +11,7 @@
       <div class="input-group-append">
          <button type="submit"
             asp-controller="Notes"
-            asp-action="Index"            
+            asp-action="Index"
             class="btn btn-outline-secondary">
             SEARCH
          </button>
