@@ -1,6 +1,5 @@
 @using Microsoft.AspNetCore.Identity
-@using Microsoft.AspNetCore.Identity.EntityFrameworkCore
-@using NoteWorx.Infrastructure.Data.Models
+@using NoteWorx.Identity.Data.Models
 
 @inject UserManager<AppUser> UserManager
 
@@ -25,40 +24,40 @@
 </head>
 <body>
     <div class="container">
-       <nav class="navbar navbar-expand-lg navbar-light bg-light">         
-         <a asp-controller="Notes" asp-action="Index" class="navbar-brand text-primary">
-            <i class="fa fa-book-open fa-2x mr-2 text-primary"></i>
-            NOTEWORX
-         </a>
-         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#menu">
-            <span class="navbar-toggler-icon"></span>
-         </button>
+        <nav class="navbar navbar-expand-lg navbar-light bg-light">
+            <a asp-controller="Notes" asp-action="Index" class="navbar-brand text-primary">
+                <i class="fa fa-book-open fa-2x mr-2 text-primary"></i>
+                NOTEWORX
+            </a>
+            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#menu">
+                <span class="navbar-toggler-icon"></span>
+            </button>
 
-         <div class="collapse navbar-collapse" id="menu">
-            <form asp-controller="Account" asp-action="Logout" class="my-2 ml-auto my-lg-0">                
-                  <ul class="navbar-nav">
-                     <li class="nav-item dropdown">
-                        <a class="nav-link dropdown-toggle" href="#" id="accountDropdown" role="button" data-toggle="dropdown">
-                              <div class="d-inline d-lg-none">ACCOUNT</div>
-                              <div class="d-none d-lg-inline">
-                                 <span class="fa-stack">
-                                 <i class="fas fa-circle fa-stack-2x text-primary"></i>
-                                 <i class="fas fa-user fa-stack-1x fa-inverse"></i>
-                                 </span>
-                                 @UserManager.GetUserName(User)
-                              </div>
-                              
-                        </a>
-                        <div class="dropdown-menu">
-                              <button type="submit" class="dropdown-item btn btn-link">Logout</button>
-                        </div>                        
-                     </li>
-                  </ul>
-            </form>
-         </div>
+            <div class="collapse navbar-collapse" id="menu">
+                <form asp-controller="Account" asp-action="Logout" class="my-2 ml-auto my-lg-0">
+                    <ul class="navbar-nav">
+                        <li class="nav-item dropdown">
+                            <a class="nav-link dropdown-toggle" href="#" id="accountDropdown" role="button" data-toggle="dropdown">
+                                <div class="d-inline d-lg-none">ACCOUNT</div>
+                                <div class="d-none d-lg-inline">
+                                    <span class="fa-stack">
+                                        <i class="fas fa-circle fa-stack-2x text-primary"></i>
+                                        <i class="fas fa-user fa-stack-1x fa-inverse"></i>
+                                    </span>
+                                    @UserManager.GetUserName(User)
+                                </div>
+
+                            </a>
+                            <div class="dropdown-menu">
+                                <button type="submit" class="dropdown-item btn btn-link">Logout</button>
+                            </div>
+                        </li>
+                    </ul>
+                </form>
+            </div>
         </nav>
     </div>
-    
+
     <div class="container">
         @RenderBody()
     </div>
