@@ -19,7 +19,7 @@
 </head>
 <body>
     <div class="container">
-       @RenderBody()
+        @RenderBody()
     </div>
 </body>
 </html>
