@@ -9,7 +9,7 @@
         <link rel="stylesheet" href="~/node_modules/@@fortawesome/fontawesome-free/css/all.min.css">
         <link rel="stylesheet" href="~/css/site.css">
     </environment>
-    
+
     <environment include="Production">
         <link rel="stylesheet" href="~/lib/css/bootstrap.min.css" asp-append-version="true">
         <link rel="stylesheet" href="~/lib/css/font-awesome.min.css" asp-append-version="true">
